# dymer-webserver/swagger_webserver.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DYMER API",
    "description": "Documentation of <b>Webserver</b> module."
  },
  "host": "localhost:0000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/templates/api/v1/template/deletelog/{filetype}": {
      "get": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "filetype",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/openLog/{filetype}": {
      "get": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "filetype",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/templates/api/v1/template/setlogconfig": {
      "post": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "consoleactive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/checkservice": {
      "get": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/*": {
      "get": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/deletelog/{filetype}": {
      "get": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "filetype",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorizationtk",
            "in": "header",
            "type": "string"
          },
          {
            "name": "extrainfo",
            "in": "header",
            "type": "string"
          },
          {
            "name": "referer",
            "in": "header",
            "type": "string"
          },
          {
            "name": "reqfrom",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "tkdymat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkextra",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkdym",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/openLog/{filetype}": {
      "get": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "filetype",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorizationtk",
            "in": "header",
            "type": "string"
          },
          {
            "name": "extrainfo",
            "in": "header",
            "type": "string"
          },
          {
            "name": "referer",
            "in": "header",
            "type": "string"
          },
          {
            "name": "reqfrom",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "tkdymat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkextra",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkdym",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/checkservice": {
      "get": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorizationtk",
            "in": "header",
            "type": "string"
          },
          {
            "name": "extrainfo",
            "in": "header",
            "type": "string"
          },
          {
            "name": "referer",
            "in": "header",
            "type": "string"
          },
          {
            "name": "reqfrom",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "tkdymat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkextra",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkdym",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/templates/*": {
      "get": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorizationtk",
            "in": "header",
            "type": "string"
          },
          {
            "name": "extrainfo",
            "in": "header",
            "type": "string"
          },
          {
            "name": "referer",
            "in": "header",
            "type": "string"
          },
          {
            "name": "reqfrom",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "tkdymat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkextra",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkdym",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      },
      "post": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorizationtk",
            "in": "header",
            "type": "string"
          },
          {
            "name": "extrainfo",
            "in": "header",
            "type": "string"
          },
          {
            "name": "referer",
            "in": "header",
            "type": "string"
          },
          {
            "name": "reqfrom",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "tkdymat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkextra",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkdym",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorizationtk",
            "in": "header",
            "type": "string"
          },
          {
            "name": "extrainfo",
            "in": "header",
            "type": "string"
          },
          {
            "name": "referer",
            "in": "header",
            "type": "string"
          },
          {
            "name": "reqfrom",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "tkdymat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkextra",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkdym",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorizationtk",
            "in": "header",
            "type": "string"
          },
          {
            "name": "extrainfo",
            "in": "header",
            "type": "string"
          },
          {
            "name": "referer",
            "in": "header",
            "type": "string"
          },
          {
            "name": "reqfrom",
            "in": "header",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          },
          {
            "name": "tkdymat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkextra",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tkdym",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/setlogConfig": {
      "post": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}